+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Changelog
 =========
+2.6.0
+-----
+
+* Implemented a new configuration option `redirectable_url_matcher` under `dynamic`.
+  If set to `true`, the router will try to detect and redirect between URLs with and without trailing slashes. See https://symfony.com/doc/4.1/routing.html#routing-trailing-slash-redirection.
+
 2.5.1
 -----